40播放进度条的控制

+++ app/src/main/java/com/example/himalaya/interfaces/IPlayerCallback.java
@@ -58,7 +58,7 @@ public interface IPlayerCallback {
      * @param currentProgress
      * @param total
      */
-    void onProgressChange(long currentProgress, long total);
+    void onProgressChange(int currentProgress, int total);
 
     /**
      * 广告正在加载
